+++ voicevox_engine/app/application.py
@@ -31,6 +31,7 @@ from voicevox_engine.preset.preset_manager import PresetManager
 from voicevox_engine.resource_manager import ResourceManager
 from voicevox_engine.setting.model import CorsPolicyMode
 from voicevox_engine.setting.setting_manager import SettingHandler
+from voicevox_engine.tts_pipeline.song_engine import SongEngineManager
 from voicevox_engine.tts_pipeline.tts_engine import TTSEngineManager
 from voicevox_engine.user_dict.user_dict_manager import UserDictionary
 from voicevox_engine.utility.path_utility import engine_root
@@ -39,6 +40,7 @@ from voicevox_engine.utility.runtime_utility import is_development
 
 def generate_app(
     tts_engines: TTSEngineManager,
+    song_engines: SongEngineManager,
     core_manager: CoreManager,
     setting_loader: SettingHandler,
     preset_manager: PresetManager,
@@ -85,7 +87,12 @@ def generate_app(
     )
 
     app.include_router(
-        generate_tts_pipeline_router(tts_engines, preset_manager, cancellable_engine)
+        generate_tts_pipeline_router(
+            tts_engines,
+            song_engines,
+            preset_manager,
+            cancellable_engine
+        )
     )
     app.include_router(generate_morphing_router(tts_engines, metas_store))
     app.include_router(
